+++ src/internalStream.js
@@ -1,24 +1,9 @@
 // @flow
 const EventEmitter = require('events');
 const { StreamError } = require('./errors');
+// const { strFromEvent } = require('./event');
 
-type EventDone = {| done: true |};
-type EventError  = {| error: Error, done?: false |};
-type EventValue<T>  = {| value: T, done?: false |};
-export type Event<T> = EventDone | EventError | EventValue<T>;
-
-function strFromEvent<T>(event: Event<T>): string {
-  if (event.error) {
-    return `<Error='${event.error.message}'>`;
-  }
-  else if (event.done) {
-    return '<Done>';
-  }
-  else {
-    // $FlowFixMe
-    return `<Value=${event.value}>`;
-  }
-}
+import type { Event } from './event';
 
 export type Push<T> = (Event<T>) => boolean;
 
@@ -224,6 +209,5 @@ class InternalStream<T> extends EventEmitter {
 }
 
 module.exports = {
-  InternalStream,
-  strFromEvent
+  InternalStream
 };
